requirements.txt
# Core Dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.1

# AI & LLM Integration
openai==1.35.3
langchain==0.2.5
langchain-openai==0.1.13
tiktoken==0.7.0

# Web Scraping & Data
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
firecrawl-py==0.0.16

# Database & Storage
sqlalchemy==2.0.30
supabase==2.5.0
redis==5.0.6
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1

# Payment & Billing
stripe==9.10.0

# Email & Notifications
sendgrid==6.11.0
python-telegram-bot==21.3

# Plugin System
wasmtime==21.0.0
pywasm==1.0.8

# Monitoring & Logging
sentry-sdk==2.5.1
structlog==24.2.0
prometheus-client==0.20.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx[cli]==0.27.0
faker==25.8.0

# Development Tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.26

# Deployment
gunicorn==22.0.0
docker==7.1.0
python-json-logger==2.0.7